requirements.txt
python-telegram-bot[job-queue]==21.6
SQLAlchemy==2.0.36
python-dotenv==1.0.1
pytz==2024.2
